pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "kedro-catalog"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
description = "Next-generation DataCatalog for Kedro"
dependencies = [
    "pydantic>=2.8.2",
]
authors = [
    {name = "Juan Luis Cano Rodríguez", email = "juan_luis_cano@mckinsey.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = ["version"]

[project.urls]
source = "https://github.com/astrojuanlu/kedro-catalog"
tracker = "https://github.com/astrojuanlu/kedro-catalog/issues"
documentation = "https://kedro-catalog.readthedocs.io"

[project.optional-dependencies]
test = [
    "mypy",
    "hypothesis",
    "pytest",
    "pytest-cov",
]
doc = [
    "furo",
    "myst-parser",
    "sphinx>=5",
    "sphinx-copybutton",
]

[tool.pdm.version]
source = "scm"

[tool.ruff]
show-fixes = true

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "E",  # Pycodestyle
    "W",  # Pycodestyle
    # "D",  # Pydocstyle
    "UP",  # pyupgrade
    "I",  # isort
    "PL",  # Pylint
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["D", "PLR2004"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
python_version = "3.8"
warn_redundant_casts = true
warn_unused_configs = true
pretty = true
show_error_codes = true
show_error_context = true
show_column_numbers = true

disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true

# More strict checks for library code
[[tool.mypy.overrides]]
module = "kedro_catalog"
disallow_untyped_defs = true

# Ignore certain missing imports
# [[tool.mypy.overrides]]
# module = "thirdparty.*"
# ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.uv]
dev-dependencies = [
    "ipython>=8.12.3",
    "ipykernel>=6.29.5",
    "kedro>=0.19.8",
    "kedro-datasets[polars]>=1.8.0",
]
